requirements_server.txt
fastapi==0.111.0
uvicorn==0.29.0
python-telegram-bot==21.2
google-generativeai==0.7.2
httpx==0.27.0
python-dotenv==1.0.1
websockets==12.0
